Create PublicationCodec module
